+++ src/MinEdLauncher/Settings.fs
@@ -31,7 +31,8 @@ let defaults =
       AdditionalProducts = List.empty
       DryRun = false
       ShutdownTimeout = TimeSpan.FromSeconds(10)
-      CacheDir = "" }
+      CacheDir = ""
+      GameStartDelay = TimeSpan.Zero }
     
 [<RequireQualifiedAccess>]
 type FrontierCredResult = Found of string * string * string option | NotFound of string | UnexpectedFormat of string | Error of string
@@ -176,7 +177,9 @@ type Config =
       AdditionalProducts: AuthorizedProduct list
       [<DefaultValue("10")>]
       ShutdownTimeout: int
-      CacheDir: string option }
+      CacheDir: string option
+      [<DefaultValue("0")>]
+      GameStartDelay: int }
 let parseConfig fileName =
     let configRoot = ConfigurationBuilder()
                         .AddJsonFile(fileName, false)
@@ -262,5 +265,6 @@ let getSettings args appDir fileConfig = task {
                                                           AdditionalProducts = fileConfig.AdditionalProducts
                                                           ShutdownTimeout = TimeSpan.FromSeconds(fileConfig.ShutdownTimeout)
                                                           CacheDir = fileConfig.CacheDir |> Option.defaultValue Environment.cacheDir
+                                                          GameStartDelay = TimeSpan.FromSeconds(fileConfig.GameStartDelay) 
                                            })
 }